Initialize the struct commands centre
The commands_centre struct is an integral part of the shell, as it keeps track
of the commands and their various meta information. At a small cost of memory,
I preferred to use this to avoid making multiple calls for extraction and
command state management.
It's efficieny is pegged in it being initialized in main and being reused
everywhere the commands and their meta information are required.

Signed-off-by: Josphat Kahihia <josphatkahihia@gmail.com>