Add execute functionality (v1)
This commit will add the ability to execute binaries and shell scripts
(using other shells only - for the moment). It can only take in commands
without their args. That is for the upcoming v0.2.0.

Having completed the requirements for v0.1.0, this commit will be
tagged as v0.1.0-alpha to polish up the error messages, various
code style issues before being released as v0.1.0.

The release of v0.1.0 will open the door for creating and addressing
issues on github, to propel development further.

Signed-off-by: Josphat Kahihia <josphatkahihia@gmail.com>